requirements-lambda.txt
fastapi==0.110.0
pydantic==2.11.4
pymongo==4.13.0
python-dotenv==1.1.0
mangum==0.17.0
groq==0.25.0
langchain-core==0.3.61
langchain-groq==0.3.2
uvicorn==0.29.0
